🚧🎨🔥Refactored views out of propertyDetailsScreen.js
Moved images, loading screens (with ui update incl. translation), and main information of property to separate files
