+++ app/Domits/src/features/translation/store/en.json
@@ -14,6 +14,7 @@
     "amenities": "Amenities",
     "show all amenities": "Show all amenities",
     "hosted by": "Hosted by",
-    "what does this place have to offer?": "What does this place have to offer?"
+    "what does this place have to offer?": "What does this place have to offer?",
+    "loading": "Loading"
   }
 }
