+++ app/Domits/src/features/translation/store/nl.json
@@ -14,6 +14,7 @@
     "amenities": "Voorzieningen",
     "show all amenities": "Toon alle voorzieningen",
     "hosted by": "Aangeboden door",
-    "what does this place have to offer?": "Wat voor voorzieningen biedt deze plek?"
+    "what does this place have to offer?": "Wat voor voorzieningen biedt deze plek?",
+    "loading": "Aan het laden van"
   }
 }
