+++ app/Domits/src/screens/propertyDetailsScreen/screens/propertyDetailsScreen.js
@@ -1,25 +1,19 @@
 import React, {useEffect, useState} from 'react';
-import {
-    ActivityIndicator,
-    Dimensions,
-    Image,
-    ScrollView,
-    Text,
-    ToastAndroid,
-    TouchableOpacity,
-    View,
-} from 'react-native';
+import {ScrollView, Text, ToastAndroid, TouchableOpacity, View,} from 'react-native';
 import {SafeAreaView} from 'react-native-safe-area-context';
-import featureIcons from '../../../ui-components/FeatureIcons';
 import FetchOwnerData from '../../../features/search/hooks/FetchOwnerData';
 import FetchAccommodation from '../../../features/search/hooks/FetchAccommodation';
 import Ionicons from "react-native-vector-icons/Ionicons";
 import {styles} from "../styles/propertyDetailsStyles";
 import AmenitiesPopup from "../components/AmenitiesPopup";
-import SelectBookingDateCalendar from "../components/SelectBookingDateCalendar";
-import BookingPopup from "../components/BookingPopup";
+import SelectBookingDatesCalendarView from "../views/SelectBookingDatesCalendarView";
 import TranslatedText from "../../../features/translation/components/TranslatedText";
 import CalculateNumberOfNights from "../../../features/bookingengine/utils/CalculateNumberOfNights";
+import RenderAmenities from "../hooks/RenderAmenities";
+import PropertyImagesView from "../views/PropertyImagesView";
+import LoadingScreen from "../../loadingscreen/screens/LoadingScreen";
+import PropertyMainDetailsView from "../views/PropertyMainDetailsView";
+import PropertyDetailsOwnerView from "../views/PropertyDetailsOwnerView";
 
 const PropertyDetailsScreen = ({route, navigation}) => {
     const accommodationId = route.params.accommodation.id;
@@ -28,23 +22,12 @@ const PropertyDetailsScreen = ({route, navigation}) => {
     const [loading, setLoading] = useState(true);
     const [images, setImages] = useState([]);
     const [showAmenitiesModal, setShowAmenitiesModal] = useState(false);
-    const [showBookingModal, setShowBookingModal] = useState(false);
-    const [currentPage, setCurrentPage] = useState(0);
-    // Dynamically calculate image width based on screen width
-    const imageWidth = Dimensions.get('window').width;
     const [firstSelectedDate, setFirstSelectedDate] = useState(null);
     const [lastSelectedDate, setLastSelectedDate] = useState(null);
     const [amountOfNights, setAmountOfNights] = useState(null);
     const [costOfNights, setCostOfNights] = useState(null);
     const [totalCost, setTotalCost] = useState(null);
 
-    const handleFirstDateSelected = (date) => {
-        setFirstSelectedDate(date);
-    };
-
-    const handleLastDateSelected = (date) => {
-        setLastSelectedDate(date);
-    }
     /**
      * Fetch current accommodation data.
      */
@@ -95,6 +78,14 @@ const PropertyDetailsScreen = ({route, navigation}) => {
         navigation.navigate('HomeScreen');
     };
 
+    const handleFirstDateSelected = (date) => {
+        setFirstSelectedDate(date);
+    };
+
+    const handleLastDateSelected = (date) => {
+        setLastSelectedDate(date);
+    }
+
     const handleOnBookPress = () => {
         if (!firstSelectedDate || !lastSelectedDate) {
             ToastAndroid.show("Please select a start and end date", ToastAndroid.SHORT)
@@ -107,106 +98,13 @@ const PropertyDetailsScreen = ({route, navigation}) => {
             });
         }
     };
-    const handleScroll = event => {
-        const page = Math.round(
-            event.nativeEvent.contentOffset.x /
-            event.nativeEvent.layoutMeasurement.width,
-        );
-        setCurrentPage(page);
-    };
-
-    const toggleBookingModal = () => {
-        setShowBookingModal(!showBookingModal);
-    }
 
     const toggleAmenitiesModal = () => {
         setShowAmenitiesModal(!showAmenitiesModal);
     };
 
-    const renderAmenities = () => {
-        const allAmenities = parsedAccommodation.Features || {};
-        const categoriesToShow = Object.keys(allAmenities)
-            .filter(category => allAmenities[category].length > 0)
-            .slice(0, 3);
-
-        return categoriesToShow.map((category, categoryIndex) => {
-            const items = allAmenities[category].slice(0, 5);
-
-            return (
-                <View key={categoryIndex}>
-                    <Text style={styles.categorySubTitle}>{category}</Text>
-                    <View style={styles.subCategoryDivider}/>
-
-                    {items.map((item, itemIndex) => (
-                        <View key={itemIndex} style={styles.featureIconItem}>
-                            {featureIcons[item] ? (
-                                <View style={styles.featureIcon}>{featureIcons[item]}</View>
-                            ) : null}
-                            <Text style={styles.featureText}>{item}</Text>
-                        </View>
-                    ))}
-                </View>
-            );
-        });
-    };
-
-    const renderDateRange = () => {
-        const dateRanges = parsedAccommodation.DateRanges || [];
-
-        if (dateRanges.length === 0) {
-            return null;
-        }
-
-        const sortedRanges = dateRanges.sort(
-            (a, b) => new Date(a.startDate) - new Date(b.startDate),
-        );
-
-        const earliestDate = new Date(sortedRanges[0].startDate);
-        const latestDate = new Date(sortedRanges[sortedRanges.length - 1].endDate);
-
-        const formatDate = date => {
-            const day = date.toLocaleDateString('en-US', {
-                day: 'numeric',
-                timeZone: 'UTC',
-            });
-            const month = date.toLocaleDateString('en-US', {
-                month: 'long',
-                timeZone: 'UTC',
-            });
-            return `${day} ${month}`;
-        };
-
-        return (
-            <View>
-                <Text>
-                    {formatDate(earliestDate)} - {formatDate(latestDate)}
-                </Text>
-            </View>
-        );
-    };
-
-    const renderDotIndicator = () => {
-        return (
-            <View style={styles.dotContainer}>
-                {images.map((_, index) => (
-                    <View
-                        key={index}
-                        style={[
-                            styles.dot,
-                            currentPage === index ? styles.activeDot : styles.inactiveDot,
-                        ]}
-                    />
-                ))}
-            </View>
-        );
-    };
-
     if (loading) {
-        return (
-            <View style={styles.loaderContainer}>
-                <ActivityIndicator size="large" color="#0000ff"/>
-            </View>
-        );
+        return <LoadingScreen loadingName={"property"}/>
     }
 
     return (
@@ -223,83 +121,22 @@ const PropertyDetailsScreen = ({route, navigation}) => {
                     <Text style={styles.titleText}>{parsedAccommodation.Title}</Text>
                 </View>
 
-                <ScrollView style={{flex: 1}}>
-                    <ScrollView
-                        horizontal={true}
-                        contentContainerStyle={styles.imageContainer}
-                        pagingEnabled={true}
-                        onScroll={handleScroll}
-                        scrollEventThrottle={100}>
-                        {Object.entries(images).map(([key, url]) => (
-                            <View key={key}>
-                                <Image
-                                    source={{uri: url.uri}}
-                                    style={[styles.image, {width: imageWidth}]}
-                                />
-                            </View>
-                        ))}
-                    </ScrollView>
-                    <View style={styles.counterContainer}>{renderDotIndicator()}</View>
+                <ScrollView style={styles.contentContainer}>
+                    <PropertyImagesView
+                        images={images}
+                    />
 
-                    <View style={styles.mainContentContainer}>
-                        <View style={styles.mainInfoContainer}>
-                            <Text style={styles.subtitleText}>
-                                {parsedAccommodation.Subtitle.trim()}
-                            </Text>
-                            <Text style={styles.costPerNightText}>
-                                €{Number(parsedAccommodation.Rent).toFixed(2)} {" "}
-                                <TranslatedText textToTranslate={"per night"}/>
-                            </Text>
-                            <Text>{renderDateRange()}</Text>
-                        </View>
+                    <View style={styles.propertyDetailsContainer}>
+                        <PropertyMainDetailsView
+                            property={parsedAccommodation}
+                        />
 
-                        <View style={styles.borderContainer}>
-                            <View style={styles.mainAmenityContainer}>
-                                <View style={styles.featureIcon}>
-                                    {featureIcons["Multiple guests"]}
-                                </View>
-                                <Text style={styles.mainAmenitiesText}>
-                                    {parsedAccommodation.GuestAmount} <TranslatedText textToTranslate={"guests"}/>
-                                </Text>
-                            </View>
-                            <View style={styles.mainAmenityContainer}>
-                                <View style={styles.featureIcon}>
-                                    {featureIcons["Bedroom"]}
-                                </View>
-                                <Text style={styles.mainAmenitiesText}>
-                                    {parsedAccommodation.bedrooms || 0} <TranslatedText textToTranslate={"bedrooms"}/>
-                                </Text>
-                            </View>
-                            <View style={styles.mainAmenityContainer}>
-                                <View style={styles.featureIcon}>
-                                    {featureIcons["Bed"]}
-                                </View>
-                                <Text style={styles.mainAmenitiesText}>
-                                    {parsedAccommodation.Beds} <TranslatedText textToTranslate={"beds"}/>
-                                </Text>
-                            </View>
-                            <View style={styles.mainAmenityContainer}>
-                                <View style={styles.featureIcon}>
-                                    {featureIcons["Bathroom"]}
-                                </View>
-                                <Text style={styles.mainAmenitiesText}>
-                                    {parsedAccommodation.Bathrooms} <TranslatedText textToTranslate={"bathrooms"}/>
-                                </Text>
-                            </View>
-                        </View>
-
-
-                        <View style={styles.descriptionContainer}>
-                            <Text style={styles.descriptionText}>
-                                {parsedAccommodation.Description.trim()}
-                            </Text>
-                        </View>
-
-                        <SelectBookingDateCalendar
+                        <SelectBookingDatesCalendarView
                             onFirstDateSelected={handleFirstDateSelected}
                             onLastDateSelected={handleLastDateSelected}
                             property={parsedAccommodation}
-                        ></SelectBookingDateCalendar>
+                            handleOnBookPress={handleOnBookPress}
+                        ></SelectBookingDatesCalendarView>
 
                         <View>
                             <Text>Nights: €{Number(parsedAccommodation.Rent * amountOfNights).toFixed(2)}</Text>
@@ -316,20 +153,13 @@ const PropertyDetailsScreen = ({route, navigation}) => {
                                     <Ionicons name={'arrow-forward-circle-outline'} size={24}
                                               style={styles.bookingButtonIcon}/>
                                 </TouchableOpacity>
-                                {showBookingModal && (
-                                    <BookingPopup
-                                        onClose={toggleBookingModal}
-                                        startDate={firstSelectedDate}
-                                        endDate={lastSelectedDate}
-                                    />
-                                )}
                             </View>
                         </View>
 
                         <View style={styles.categoryDivider}/>
 
                         <Text style={styles.categoryTitle}><TranslatedText textToTranslate={"amenities"}/></Text>
-                        <View style={styles.amenities}>{renderAmenities()}</View>
+                        <View style={styles.amenities}>{RenderAmenities(parsedAccommodation)}</View>
                         <View>
                             <TouchableOpacity
                                 onPress={toggleAmenitiesModal}
